All 3 detection models implemented

+++ Backend/cv_engine/schemas.py
@@ -4,5 +4,5 @@ from typing import Optional, Tuple
 class DetectionEvent(BaseModel):
     type: str                # "choking", "bleeding", "unresponsive"
     confidence: float
-    coords: Optional[Tuple[int, int]] = None  # (x, y) coordinates on frame
+    coords: Optional[Tuple[float, float]] = None  # (x, y) coordinates on frame
     frame_id: Optional[int] = None
